Cambios realizados: -Reubicacion de providers de la carpeta de components a la carpeta de providers -Documentacion de los componentes de la calculadora -Eliminado de comentarios innecesarios y exportaciones no utilizadas
