--- src/calculadora/components/LatexInputs.jsx
@@ -1,6 +1,0 @@
-import { useState } from "react";
-
-
-export const LatexInputs = () => {
-
-};
