+++ src/calculadora/components/index.js
@@ -6,7 +6,5 @@ export * from "./ImageModal"
 export * from "./ImageSlider"
 export * from "./ModalTutorial"
 export * from "./LinkComponent"
-export * from "./LoaderProvider"
 export * from "./ChatComponent"
 export * from "./InputMessageComponent"
-export * from "./ChatProvider"
